GPU rendering now uses 64bit image map pixel indices (i.e. removed the limit of 8 * 4GB of scene image maps)

+++ include/slg/engines/pathoclbase/compiledscene.h
@@ -173,7 +173,7 @@ private:
 	Scene *scene;
 	const PathTracer *pathTracer;
 
-	u_int maxMemPageSize;
+	size_t maxMemPageSize;
 	boost::unordered_set<std::string> enabledCode;
 };
